@@ -3,20 +3,18 @@
 
 #include <GLES2/gl2.h>
 
-#include "../gamestate.h"
-#include "../sceneManager/sceneManager.h"
+#include "../viewController/viewController.h"
 
 class Renderer {
 public:
-    Renderer(GameState* g, SceneManager* s);
+    Renderer(ViewController* v);
     ~Renderer();
 
     void setClearColor(float r, float g, float b, float a);
     void resize(int w, int h);
     void render();
 private:
-    GameState* gameState;
-    SceneManager* sceneManager;
+    ViewController* viewController;
 };
 
 #endif
